Make `test/marshalling_fields.cpp` write to stdout
Just to ease pollution of the system journal when testing.

For some godforsaken reason, checking for equality doesn't work when
comparing the output string to the ideal output (even when its own
output is copied from a debugger like GDB). So this has to be actually
written to a file and checked manually in order to verify that it works.
If someone has a better solution, though, please do not hesitate to
notify me and/or offer a fix.

- Dr. Bluefall.